requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
Pillow>=10.2.0
django-filter==23.5
django-celery-beat==2.5.0
django-celery-results>=2.5.0

